New classifiers from CBPF team
