+++ fink_science/cats/requirements.txt
@@ -1,0 +1,3 @@
+tensorflow==2.8.0
+numpy
+tensorflow-addons
