--- fink_science/cbpf_classifier/README.md
@@ -1,2 +1,0 @@
-# Hierarchical Classifier
-
