--- fink_science/cbpf_classifier/requirements.txt
@@ -1,3 +1,0 @@
-tensorflow==2.8.0
-numpy
-tensorflow-addons
